+++ super_repo/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+"""Example.
+
+Examples for the calculater functionality to demonstrate TDD.
+
+SPDX-FileCopyrightText: Jonas Huber (jh-RLI) © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Ludwig Hülk (Ludee) © Reiner Lemoine Institut
+SPDX-License-Identifier: MIT
+"""