+++ super_repo/example_calculator.py
@@ -2,12 +2,10 @@
 
 """Example.
 
-Example implementation of test for the calculater functionality
-to demonstrate TDD.
+Examples for the calculater functionality to demonstrate TDD.
 
-
-SPDX-FileCopyrightText: 2023 Jonas Huber <@jh-RLI>
-SPDX-FileCopyrightText: © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Jonas Huber (jh-RLI) © Reiner Lemoine Institut
+SPDX-FileCopyrightText: Ludwig Hülk (Ludee) © Reiner Lemoine Institut
 SPDX-License-Identifier: MIT
 """